\name{Association test}
\alias{AStest}
\title{Association between genotype and covariates}
\usage{
AStest(y=y, X=X, C=C)
y represent phenotype,
X represent genotype,
C represent covariates
}
\description{
This is a function for test the correlation between genotype and covariates.
}
\examples{
AStest(y=myY,X=myGD,C=myC)
}
